requirements.txt
picamera2==0.3.12
paho-mqtt==1.6.1
python-dotenv==1.0.0
types-paho-mqtt==1.6.0.20240106 